fortis_bot_export/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fortis-bot"
version = "2.0.0"
description = "Trading bot implementing the Fortis Trading System"
requires-python = ">=3.11"

dependencies = [
    "python-binance>=1.0.19",
    "supabase>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "ta>=0.10.2",
    "python-telegram-bot>=20.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-asyncio>=0.21.0", "black>=23.0.0"]
